Add files via upload
These are the tasks which I have built during my internship in CodSoft. And I assure you that all these models are built on my own with reference to other models.
I have also uploaded these codes in another github profile of mine. So if in case you find that these codes are similar to the github link i have provided below , please ignore it.
https://github.com/S-Prathyush